feat: implementação web API

+++ src/main/java/io/github/libraryapi/service/AutorService.java
@@ -4,25 +4,41 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import io.github.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.libraryapi.model.Autor;
 import io.github.libraryapi.repository.AutorRepository;
+import io.github.libraryapi.repository.LivroRepository;
+import io.github.libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
+    @Autowired
     private final AutorRepository repository;
 
-    public AutorService(AutorRepository repository) {
-        this.repository = repository;
+    @Autowired
+    private final AutorValidator validator;
+
+    @Autowired
+    private final LivroRepository livroRepository;
+
+    public Autor salvar(Autor autor) {
+        validator.validar(autor);
+        return repository.save(autor);
     }
 
     public void atualizar(Autor autor) {
         if (autor.getId() == null) {
             throw new IllegalArgumentException("Para atualizar é necessário que o Autor esteja salvo na base");
         }
+        validator.validar(autor);
         repository.save(autor);
+
     }
 
     public Optional<Autor> obterPorId(UUID id) {
@@ -30,6 +46,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir, Autor possui livros cadastrados");
+        }
         repository.delete(autor);
     }
 
@@ -47,6 +66,12 @@ public class AutorService {
         }
 
         return repository.findAll();
+
+    }
+
+    public boolean possuiLivro(Autor autor) {
+
+        return livroRepository.existsByAutor(autor);
 
     }
 }
